backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.3.0
langchain-openai==0.2.0
langgraph==0.2.0
openai==1.54.0
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
